+++ input/drivers/qnx_input.c
@@ -761,7 +761,7 @@ static int16_t qnx_pointer_input_state(qnx_input_t *qnx,
    return 0;
 }
 
-static int16_t qnx_input_state(
+static int32_t qnx_input_state(
       void *data,
       const input_device_driver_t *joypad,
       const input_device_driver_t *sec_joypad,
@@ -781,7 +781,7 @@ static int16_t qnx_input_state(
          if (id == RETRO_DEVICE_ID_JOYPAD_MASK)
          {
             unsigned i;
-            int16_t ret = 0;
+            int32_t ret = 0;
 
             if (!keyboard_mapping_blocked)
             {
