+++ input/drivers/winraw_input.c
@@ -723,7 +723,7 @@ static unsigned winraw_retro_id_to_rarch(unsigned id)
    return 0;
 }
 
-static int16_t winraw_input_state(
+static int32_t winraw_input_state(
       void *data,
       const input_device_driver_t *joypad,
       const input_device_driver_t *sec_joypad,
@@ -735,7 +735,7 @@ static int16_t winraw_input_state(
       unsigned idx,
       unsigned id)
 {
-   int16_t ret           = 0;
+   int32_t ret           = 0;
    settings_t *settings  = NULL;
    winraw_mouse_t *mouse = NULL;
    winraw_input_t *wr    = (winraw_input_t*)data;
