+++ input/input_types.h
@@ -57,7 +57,7 @@ struct turbo_buttons
 {
    int32_t turbo_pressed[MAX_USERS];
    unsigned count;
-   uint16_t enable[MAX_USERS];
+   uint32_t enable[MAX_USERS];
    bool frame_enable[MAX_USERS];
    bool mode1_enable[MAX_USERS];
 };
